Rename RiskScoreClassificationValidationError to GeneralValidationError
