+++ services/distribution/src/test/java/app/coronawarn/server/services/distribution/assembly/appconfig/validation/RiskScoreClassificationValidatorTest.java
@@ -20,14 +20,14 @@
 
 package app.coronawarn.server.services.distribution.assembly.appconfig.validation;
 
-import static app.coronawarn.server.services.distribution.assembly.appconfig.validation.RiskScoreClassificationValidationError.ErrorType.INVALID_PARTITIONING;
-import static app.coronawarn.server.services.distribution.assembly.appconfig.validation.RiskScoreClassificationValidationError.ErrorType.MIN_GREATER_THAN_MAX;
+import static app.coronawarn.server.services.distribution.assembly.appconfig.validation.GeneralValidationError.ErrorType.INVALID_PARTITIONING;
+import static app.coronawarn.server.services.distribution.assembly.appconfig.validation.GeneralValidationError.ErrorType.MIN_GREATER_THAN_MAX;
 import static java.util.Arrays.asList;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import app.coronawarn.server.common.protocols.internal.RiskScoreClass;
 import app.coronawarn.server.common.protocols.internal.RiskScoreClassification;
-import app.coronawarn.server.services.distribution.assembly.appconfig.validation.RiskScoreClassificationValidationError.ErrorType;
+import app.coronawarn.server.services.distribution.assembly.appconfig.validation.GeneralValidationError.ErrorType;
 import java.util.Arrays;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
@@ -148,8 +148,8 @@ class RiskScoreClassificationValidatorTest {
     ).map(Arguments::of);
   }
 
-  private static RiskScoreClassificationValidationError buildError(String parameter, Object value, ErrorType reason) {
-    return new RiskScoreClassificationValidationError(parameter, value, reason);
+  private static GeneralValidationError buildError(String parameter, Object value, ErrorType reason) {
+    return new GeneralValidationError(parameter, value, reason);
   }
 
   private static RiskScoreClassificationValidator buildValidator(RiskScoreClass... riskScoreClasses) {
@@ -164,7 +164,7 @@ class RiskScoreClassificationValidatorTest {
     return RiskScoreClass.newBuilder().setLabel(label).setMin(min).setMax(max).setUrl(url).build();
   }
 
-  private static ValidationResult buildExpectedResult(RiskScoreClassificationValidationError... errors) {
+  public static ValidationResult buildExpectedResult(GeneralValidationError... errors) {
     var validationResult = new ValidationResult();
     Arrays.stream(errors).forEach(validationResult::add);
     return validationResult;
